requirements.txt
# 딥러닝/ML 핵심 라이브러리
# 참고: CUDA 12.4용 PyTorch는 별도로 설치해야 할 수 있습니다.
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
torch>=2.1.0
transformers>=4.56.1
accelerate>=0.24.0

# vLLM (고성능 추론용, Flask 통신에 사용)
# 참고: CUDA 12.4 환경에서는 소스 빌드가 필요할 수도 있습니다
# 자세한 내용: https://docs.vllm.ai/en/latest/getting_started/installation.html
vllm>=0.4.2

# 이미지 처리 라이브러리
Pillow>=10.2.0
numpy>=1.24.0

# PDF 처리
PyMuPDF>=1.23.0  # fitz 모듈로 사용

# OCR 및 레이아웃 분석
# surya-ocr은 einops, opencv-python-headless, pydantic 등의 의존성을 포함합니다
surya-ocr>=0.17.0

# 웹 프레임워크
Flask>=2.3.0

# Qwen Vision-Language 모델 유틸리티
qwen-vl-utils>=0.1.0

# 유틸리티
tqdm>=4.66.0
requests>=2.31.0
PyYAML>=6.0

# 환경 정보
# Python: 3.10.12
# CUDA: 12.4

